Comprehensive linting cleanup: black, ruff, mypy, bandit
- black: Reformatted 141 files to consistent style
- ruff: Fixed 1,706 errors (unused imports/vars, import order,
  type comparisons, undefined names, whitespace in string literals)
- bandit: Fixed all 14 medium+ security issues:
  - Replaced pickle.load with RestrictedUnpickler allowing only safe types
  - Replaced eval() with safe AST-based evaluator for bitwise expressions
  - Replaced hardcoded /tmp/ with tempfile.gettempdir()
- mypy: Fixed 600+ type errors (return annotations, Optional types,
  isinstance guards, proper casts, union type narrowing)
- Converted try/import availability checks to importlib.util.find_spec()
- Net: -251 lines, zero test regressions (2431 passing unchanged)

+++ r2morph/mutations/parallel_executor.py
@@ -311,7 +311,7 @@ class ParallelMutator:
 
         overhead_factor = 1.0 + (0.1 * effective_workers)
 
-        return effective_workers / overhead_factor
+        return float(effective_workers / overhead_factor)
 
 
 def create_parallel_executor(config: dict[str, Any] | None = None) -> ParallelMutator:
